add conversion function to primitives
