+++ pallets/credentials/src/lib.rs
@@ -6,7 +6,6 @@ mod mock;
 #[cfg(test)]
 mod tests;
 mod types;
-mod convert;
 pub mod credential;
 
 pub use pallet::*;
@@ -29,7 +28,6 @@ pub mod pallet {
 	use crate::weights::WeightInfo;
 	use crate::types::*;
 	use crate::credential::Credential;
-	use crate::convert::*;
 
 
 	#[pallet::pallet]
@@ -232,7 +230,7 @@ pub mod pallet {
 			};
 			let did_vec: Vec<&str> = did_string.split(":").collect();
 			let public_key_str = did_vec[2].trim();
-			match convert_string_to_accountid(public_key_str){
+			match node_primitives::convert2accountid::convert_string_to_accountid(public_key_str){
 				Ok(account_id) => Ok(account_id),
 				Err(e) => Err(e),
 			}
