+++ pallets/schemas/Cargo.toml
@@ -13,7 +13,6 @@ repository = "https://github.com/serv-official/serv_testnet_backend"
 targets = ["x86_64-unknown-linux-gnu"]
 
 [dependencies]
-bs58 = { version = "0.4.0", default-features = false }
 hex-literal = { version = "0.3.4", default-features = false }
 codec = { package = "parity-scale-codec", version = "3.0.0", default-features = false, features = [
 	"derive",
@@ -31,6 +30,7 @@ frame-support = { default-features = false, git = "https://github.com/paritytech
 frame-system = { default-features = false, git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v0.9.42" }
 sp-keystore = { default-features = false, optional=true, git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v0.9.42" }
 sp-application-crypto = { default-features = false, git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v0.9.42" }
+node-primitives = {default-features = false, path = "../../primitives"}
 
 [features]
 default = ["std"]
