+++ primitives/Cargo.toml
@@ -12,14 +12,18 @@ targets = ["x86_64-unknown-linux-gnu"]
 
 [dependencies]
 codec = { package = "parity-scale-codec", version = "3.0.0", default-features = false, features = ["derive"] }
-
+bs58 = { version = "0.4.0", default-features = false }
+log = { version = "0.4.14", default-features = false }
 sp-core = { git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v0.9.42", default-features = false }
 sp-runtime = { git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v0.9.42", default-features = false }
+sp-std = { default-features = false, git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v0.9.42" }
 
 [features]
 default = ["std"]
 std = [
 	"codec/std",
+	"log/std",
+	"sp-std/std",
 	"sp-core/std",
 	"sp-runtime/std",
 ]
